requirements.txt
streamlit==1.26.0
openai==0.28.1
faiss-cpu==1.7.4
moviepy==1.0.3
imageio==2.31.1
imageio-ffmpeg==0.4.7
numpy==1.24.4
tqdm==4.65.0

